# api-gateway/internal/http/docs/swagger.yaml
basePath: /
definitions:
  client.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
    type: object
  http.CreateTaskRequest:
    properties:
      description:
        example: Outline API contract and data model
        type: string
      title:
        example: Write design doc
        type: string
    type: object
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  http.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        type: string
    type: object
  http.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret123
        type: string
    type: object
  http.TaskResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  http.TokenResponse:
    properties:
      token:
        type: string
    type: object
info:
  contact: {}
  description: API Gateway proxying auth-service and task-service.
  title: Taskboard API Gateway
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login via auth-service and receive JWT.
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Login
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user via auth-service.
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/client.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Register user
      tags:
      - auth
  /tasks:
    post:
      consumes:
      - application/json
      description: Create a task via task-service.
      parameters:
      - description: Task payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create task
      tags:
      - tasks
schemes:
- http
swagger: "2.0"